Organized to React Context, Search Functionality and Components Structured Folder
